+++ dashboard/src/app/(public)/publicUser/marketplace/page.tsx
@@ -1,11 +1,10 @@
-import React from "react";
+import ProductsList from "./productListing";
 
 export default function Marketplace (){
     return(
         <div>
-            <h1>Market plaace </h1>
-            <h1>Market plaace </h1>
-            <h1>Market plaace </h1>
+            <h1>MARKET PLACE</h1>
+            <ProductsList/>
         </div>
     )
 }
