+++ dashboard/src/app/(public)/publicUser/marketplace/productListing/index.tsx
@@ -1,0 +1,14 @@
+"use client"
+import { useState, useEffect } from "react";
+
+
+
+export default function ProductsList (){
+    const [products, setProducts] = useState([]);
+
+    return(
+        <div>
+            <h1>Products List</h1>
+        </div>
+    )
+}
