+++ dashboard/src/app/api/products/route.ts
@@ -1,0 +1,30 @@
+export async function GET() {
+    const baseUrl = process.env.BASE_URL;
+
+    if (!baseUrl) {
+        return new Response('Base URL is not configured', { status: 500 });
+    }
+
+    try {
+        const response = await fetch(`${baseUrl}/api/products/`, {
+            method: 'GET',
+            headers: {
+                'Content-Type': 'application/json',
+            },
+        });
+
+        if (!response.ok) {
+            return new Response('Failed to fetch products', { status: response.status });
+        }
+
+        const products = await response.json();
+        console.log({ all: products });
+        
+        return new Response(JSON.stringify(products), {
+            status: 200,
+            headers: { 'Content-Type': 'application/json' },
+        });
+    } catch (error) {
+        return new Response('Server error', { status: 500 });
+    }
+}
